Move feature detection static functions from MainGazeTracker into Detection
